Move rule trace tests to subject test classes
